==> notebooks/01_prepare_spatial_backbone.ipynb <==
# 01 Prepare Spatial Backbone

**Project:** NORI  
**Author:** Yuseof J  
**Date:** 09/12/25

### **Purpose**
Load the raw NY State tract shapefile, filter for NYC tracts, and set crs to Long Island/NYC. 

### **Inputs**
- `data/raw/tiger_tracts_ny/tl_2025_36_tract.shp`

### **Outputs**
- `data/processed/nyc_tracts.gpkg`
  
--------------------------------------------------------------------------

### 0. Imports and Setup

# package imports
import os
import pandas as pd
import geopandas as gpd
from pathlib import Path

# specify filepaths
path_tracts_shapefile = 'data/raw/tiger_tracts_ny/tl_2025_36_tract.shp'
path_output_processed_geodata = 'data/processed/nyc_tracts.gpkg'
path_nyc_boroughs = 'data/raw/nyc_boroughs/nybbwi.shp'

# list of fips codes for nyc - used for filtering whole ny state dataset 
nyc_county_fips = ["005", "047", "061", "081", "085"]

# EPSG:2263 - this coordinate reference system is specifically used for high-accuracy mapping of nyc boroughs 
nyc_crs = 'EPSG:2263'

# ensure cwd is project root for file paths to function properly
project_root = Path(os.getcwd())            # get current directory
while not (project_root / "data").exists(): # keep moving up until in parent
    project_root = project_root.parent
os.chdir(project_root)                      # switch to parent directory

### 1. Load Data

gdf_tracts = gpd.read_file(path_tracts_shapefile)
gdf_boroughs = gpd.read_file(path_nyc_boroughs)

### 2. EDA 

gdf_tracts.columns.tolist()

gdf_tracts.head()

gdf_boroughs.head()

### 3. Data Processing and Filtering

# ensure county fips codes are properly formatted before filtering
gdf_tracts.COUNTYFP = gdf_tracts.COUNTYFP.apply(lambda x: str(x).zfill(3))

# filter for nyc boroughs
gdf_tracts_nyc = gdf_tracts[gdf_tracts.COUNTYFP.isin(nyc_county_fips)]

print("Total tracts: ", len(gdf_tracts_nyc))

# set coordinate reference system 
gdf_tracts_nyc = gdf_tracts_nyc.to_crs(nyc_crs)

Here we'll also add borough name. This will be useful for spatial CV in the ML pipeline. Tract centroid is used rather than tract boundaries to handle tracts that span borough boundaries.  

# ensure same crs before sjoin
gdf_boroughs = gdf_boroughs.to_crs(nyc_crs)

# calculate tract centroids for point-in-polygon sjoin
gdf_tracts_nyc['centroid'] = gdf_tracts_nyc.geometry.centroid

# spatial join between tracts and broughs
gdf_tracts_nyc = gpd.sjoin(
    gdf_tracts_nyc.set_geometry('centroid'),
    gdf_boroughs[['BoroCode', 'BoroName', 'geometry']],
    how='left',
    predicate='within'
)

# drop unneeded columns
gdf_tracts_nyc.drop(columns=['index_right', 'centroid'], inplace=True)

# make sure all tracts have a borough
#assert(gdf_tracts_nyc.BoroName.isna().any() == False)

One tract did not join to a borough. Upon visual inspection in QGIS, this tract belongs to Manhattan. I'll set this manually.

# find unnasigned tract
# gdf_tracts_nyc[gdf_tracts_nyc.BoroName.isna()]

# set borough manually
gdf_tracts_nyc.loc[gdf_tracts_nyc.GEOID == '36061000100', ['BoroName', 'BoroCode']] = ['Manhattan', 1]

gdf_tracts_nyc[gdf_tracts_nyc.GEOID == '36061000100']

assert(gdf_tracts_nyc.BoroName.isna().any() == False)

gdf_tracts_nyc.head()

### 4. Save Data

# export processed tract data
gdf_tracts_nyc.to_file(path_output_processed_geodata, layer="tracts")
