==> notebooks/05_prepare_osm_green_space.ipynb <==
# 02 Prepare OSM Green Space

**Project:** NORI  
**Author:** Yuseof J  
**Date:** 10/12/25  

### **Purpose**
Load the raw OSM shapefiles, select features of interest, filter for only NYC Tracts, and output data as a GeoPackage file. 

*NOTE : I keep the OSM data relatively broad/unfilitered here, including feature classes for multipl types of green spaces. This is done simply to leave room for exploration and added functionality after the first project sprint. For now, the main feature of interest for modeling purposes will be parks (i.e. publicly accessible spaces defined specifically by fclass == 'park')*

### **Inputs**
- `data/raw/osm_ny_shp/gis_osm_landuse_a_free_1.shp`
- `data/raw/osm_ny_shp/gis_osm_natural_free_1.shp`
- `data/raw/osm_ny_shp/gis_osm_natural_a_free_1.shp`
- `data/raw/osm_ny_shp/gis_osm_pois_free_1.shp`
- `data/raw/osm_ny_shp/gis_osm_pois_a_free_1.shp`

### **Outputs**
- `data/processed/greenspaces_nyc.gpkg`
  
--------------------------------------------------------------------------

### 0. Imports and Setup

# package imports
import os
import pandas as pd
import geopandas as gpd
from pathlib import Path

# specify filepaths
path_osm_ny_dir = 'data/raw/osm_ny_shp/'
path_nyc_tracts = 'data/processed/nyc_tracts.gpkg'
path_landuse_polygons = path_osm_ny_dir + 'gis_osm_landuse_a_free_1.shp'
path_natural_points = path_osm_ny_dir + 'gis_osm_natural_free_1.shp'
path_natural_polygons = path_osm_ny_dir + 'gis_osm_natural_a_free_1.shp'
path_pois_points = path_osm_ny_dir + 'gis_osm_pois_free_1.shp'
path_pois_polygons = path_osm_ny_dir + 'gis_osm_pois_a_free_1.shp'
path_output_greenspaces_nyc = 'data/processed/greenspaces_nyc.gpkg'

# EPSG:2263 - this coordinate reference system is specifically used for high-accuracy mapping of nyc boroughs 
nyc_crs = 'EPSG:2263'

# ensure cwd is project root for file paths to function properly
project_root = Path(os.getcwd())            # get current directory
while not (project_root / "data").exists(): # keep moving up until in parent
    project_root = project_root.parent
os.chdir(project_root)                      # switch to parent directory

### 1. Load Data

# nyc tracts
gdf_tracts_nyc = gpd.read_file(path_nyc_tracts, layer="tracts")

# landuse (e.g. park, commercial, residential)
gdf_landuse_polygons = gpd.read_file(path_landuse_polygons)

# natural geographic features
gdf_natural_points = gpd.read_file(path_natural_points)
gdf_natural_polygons = gpd.read_file(path_natural_polygons)

# points of interest
gdf_pois_points = gpd.read_file(path_pois_points)
gdf_pois_polygons = gpd.read_file(path_pois_polygons)

### 2. EDA 

#### Landuse

gdf_landuse_polygons.columns.tolist()

gdf_landuse_polygons.fclass.unique()

gdf_landuse_polygons.head()

#### Natural

##### -- Points -- 

gdf_natural_points.columns.tolist()

gdf_natural_points.fclass.unique() # which feature classes are present

gdf_natural_points.head()

##### -- Polygons -- 

gdf_natural_polygons.columns.tolist()

gdf_natural_polygons.fclass.unique() # which feature classes are present

gdf_natural_polygons.head()

#### Points Of Interest

##### -- Points -- 

gdf_pois_points.columns.tolist()

gdf_pois_points.fclass.unique() # which feature classes are present

gdf_pois_points.head()

##### -- Polygons -- 

gdf_pois_polygons.columns.tolist()

gdf_pois_polygons.fclass.unique() # which feature classes are present

gdf_pois_polygons.head()

### 3. Data Processing and Filtering

##### Filter OSM layer's for green features

# filter landuse for greenspace features
landuse_features_of_interest = ['forest',
                                 'grass',
                                 'scrub',
                                 'nature_reserve',
                                 'heath',
                                 'park',
                                 'meadow',
                                 'allotments']

gdf_landuse_polygons = gdf_landuse_polygons[gdf_landuse_polygons.fclass.isin(landuse_features_of_interest)]

# filter natural for greenspace features (pints and polygons)
natural_features_of_interest_points = ['tree', 
                                        'spring', 
                                        'beach']

natural_features_of_interest_polygons = ['spring', 
                                        'beach']

gdf_natural_points = gdf_natural_points[gdf_natural_points.fclass.isin(natural_features_of_interest_points)]
gdf_natural_polygons = gdf_natural_polygons[gdf_natural_polygons.fclass.isin(natural_features_of_interest_polygons)]

# ensure no additional features exist in the polygon layer so that the list below can be used to 
# filter both the polygon and points pois gdf's
assert((set(gdf_pois_polygons.fclass.unique()) - set(gdf_pois_points.fclass.unique())) == set())

pois_features_of_interest = ['park',                                    
                             'playground',
                             'picnic_site',
                             'camp_site',
                             ]

gdf_pois_points = gdf_pois_points[gdf_pois_points.fclass.isin(pois_features_of_interest)]
gdf_pois_polygons = gdf_pois_polygons[gdf_pois_polygons.fclass.isin(pois_features_of_interest)]

# concatenate point features and convert to nyc crs
gdf_green_points = pd.concat([gdf_natural_points, gdf_pois_points], ignore_index=True)
gdf_green_points = gdf_green_points.to_crs(nyc_crs)

# concatenate polygon features and convert to nyc crs
gdf_green_polygons = pd.concat([gdf_landuse_polygons, gdf_natural_polygons, gdf_pois_polygons], ignore_index=True)
gdf_green_polygons = gdf_green_polygons.to_crs(nyc_crs)

gdf_tracts_nyc.head()

##### Filter only NYC features from overall NY State features

# select only polygons which intersect with nyc tracts
# union_all() basically creates a combined geometry of all the individual tract geometries, making calculations less complex
# only polygons that fall completely outside of nyc tracts are excluded
gdf_green_nyc_polygons = gdf_green_polygons[
    gdf_green_polygons.intersects(gdf_tracts_nyc.geometry.union_all())
].copy()

# select only points which fall within nyc tracts, similar to function above, but for points instead of polygons
gdf_green_nyc_points = gdf_green_points[
    gdf_green_points.intersects(gdf_tracts_nyc.geometry.union_all())
].copy()

### 4. Save Data

# save points and polygons as two layers in the same geopackage
gdf_green_nyc_polygons.to_file(path_output_greenspaces_nyc, layer="green_polygons")
gdf_green_nyc_points.to_file(path_output_greenspaces_nyc, layer="green_points")
