==> notebooks/06_prepare_metro_data.ipynb <==
# 02 Prepare CDC Places

**Project:** NORI  
**Author:** Yuseof J  
**Date:** 10/12/25  

### **Purpose**
Load the raw CDC PLACES csv, filter for NYC tracts, and output data as parquet file. 

### **Inputs**
- `data/raw/cdc_places.csv`

### **Outputs**
- `data/processed/cdc_places_nyc.parquet`
  
--------------------------------------------------------------------------

### 0. Imports and Setup

# package imports
import os
import pandas as pd
import geopandas as gpd
from pathlib import Path

# specify filepaths
path_cdc_places = 'data/raw/cdc_places.csv'
path_nyc_tracts = 'data/processed/nyc_tracts.gpkg'
path_output_processed_data = 'data/processed/cdc_places_nyc.parquet'

# ensure cwd is project root for file paths to function properly
project_root = Path(os.getcwd())            # get current directory
while not (project_root / "data").exists(): # keep moving up until in parent
    project_root = project_root.parent
os.chdir(project_root)                      # switch to parent directory

### 1. Load Data

df_cdc = pd.read_csv(path_cdc_places)

gdf_tracts_nyc = gpd.read_file(path_nyc_tracts, layer="tracts")

### 2. EDA 

df_cdc.columns.tolist()

df_cdc.head()

### 3. Data Filtering / Cleaning

# ensure matching dtypes for filtering
nyc_tracts_fips = gdf_tracts_nyc.GEOID.astype(int)
df_cdc.TractFIPS = df_cdc.TractFIPS.astype(int)

# filter cdc places for nyc tracts
df_cdc_nyc = df_cdc[df_cdc.TractFIPS.isin(nyc_tracts_fips)]

# report number of matched tracts
percent_matched = int((df_cdc_nyc.TractFIPS.nunique()/df_tracts_nyc.GEOID.nunique()) * 100)
print(f"Found places data for {df_cdc_nyc.TractFIPS.nunique()} / {df_tracts_nyc.GEOID.nunique()} ({percent_matched}%) of nyc tracts")

### 4. Save Data

df_cdc_nyc.to_parquet(path_output_processed_data)
