==> notebooks/14_health_vs_mobility_shap_comparison.ipynb <==
The purpose of this analysis is to compare the behavior of features between tract-level health and long-run economic mobility predictions. Specifically, this analysis will take a look at which features were important to both models, and how their values impacted each model's predictions similarly or differently.

SHAP Feature Comparison Summary

| Feature | Health SHAP | Mobility SHAP | Interpretation |
|----------|----------|----------|----------|
| pct_higher_ed       | Strong, Negative, Linear   | Weak, Positive to a point   | Matters more for health   |
| median_gross_rent   | Threshold, Negative   | Threshold, Positive   | Interacts with income / poverty  |
| pct_age_65_plus     | Negative, Linear, Weaker   | Positive, Linear, Stronger   | Matters for both   |


In-Depth Summary

side by side dependence plots and beeswarm plots?

Project Hypothesis Questions

**Do green features matter for health and economic mobility outcome predictions?**

For both the health and economic mobility models, we observe vertical clustering for outcome predictions at lower park accessibility values. This indicates that green accessibility alone may not adequately account for variation in health and economic mobility outcomes.

Conclusion

Overall, the SHAP analysis suggests that some of the most important drivers of both health and economic mobility outcomes relate more to demographic and socioeconomic characteristics of tracts.

The important thing to note, however, is that both model's are still in their first stage. With their current feature sets, they do not very robustly explain the variation in health and economic mobility outcomes among tracts. For this reason, we cannot make strong generalizations as of yet, but we can start to get an idea of which kinds of features matter and which do not.
